v0.1.16 - Addition to method `mpmToolbox.gui.score.ScoreDisplayPanel.paintComponent()`. Performance squares in the score will now contain a symbol indicating the type of the performance instruction. - Dragging a note in the piano roll is now only possible when it is an audio alignment. Performance renderings no longer allow note dragging but should be altered by performance instructions. - Added another panel to the audio analysis tab, the tempomap panel.   - New classes were added, `mpmToolbox.gui.audio.TempoMapPanel` and `mpmToolbox.gui.audio.utilities.CursorPositions`.   - The tempomap panel visualizes its data on the symbolic time axis. Class `CursorPositions` does the corresponding conversions to place playback and mouse cursor in all panels at their correct position.   - Further conversion methods hav also been added to classes `mpmToolbox.projectData.alignment.Alignment` and `mpmToolbox.projectData.alignment.Part`. Class `mpmToolbox.projectData.alignment.Note` is also slightly expanded.

+++ src/mpmToolbox/gui/syncPlayer/SyncPlayer.java
@@ -540,7 +540,7 @@ public class SyncPlayer extends WebPanel {
      * to relate the sample position to a relative playtime position
      * @param samplePosition the sample position to start playback
      */
-    public synchronized void triggerPlayback(int samplePosition) {
+    public synchronized void triggerPlayback(long samplePosition) {
         Audio audio = this.getSelectedAudio();
         if (audio == null)
             return;
